==> code/0. packages_used.R <==
install.packages('rJava')
install.packages('multilinguer')
install.packages(c('hash', 'tau', 'Sejong', 'RSQLite', 'devtools'), type = "binary")
install.packages('remotes')
install.packages('tm')
install.packages('wordcloud2')
install.packages('Sejong')
install.packages('stringr')
install.packages('arules')
install.packages('igraph')
install.packages('httr')
install.packages('rvest')
install.packages('tidytext')
install.packages('tidyverse')
install.packages('dplyr')
install.packages('tibble')
install.packages('readr')
install.packages('corrgram')
install.packages('corrplot')
install.packages('ggraph')

installremotes::install_github('haven-jeon/KoNLP', upgrade = "never", INSTALL_opts=c("--no-multiarch"))

library(rJava)
library(multilinguer)
library(hash)
library(tau)
library(RSQLite)
library(devtools)
library(remotes)
library(tm)
library(wordcloud2)
library(Sejong)
library(stringr)
library(readr)
library(arules)
library(igraph)
library(httr)
library(rvest)
library(tidytext)
library(dplyr)
library(tibble)
library(readr)
library(corrgram)
library(corrplot)
library(tidyverse)
library(ggraph)
